Add STL meshes (#27)
* adding real stl; still need collision

* add temporary offsets until we get cb/cb

* added collision

* fill holes in collision model

* change buoy name to mbari_wec; move meshes into models/mbari_wec

* fixing mesh uri
