--- buoy_description/models/buoy/model.config
@@ -1,12 +1,0 @@
-<?xml version="1.0"?>
-<model>
-  <name>Buoy</name>
-  <version>1.0</version>
-  <sdf version="1.6">model.sdf</sdf>
-
-  <author>
-  </author>
-
-  <description>
-  </description>
-</model>
