+++ buoy_description/models/mbari_wec/model.config
@@ -1,0 +1,12 @@
+<?xml version="1.0"?>
+<model>
+  <name>MBARI_WEC</name>
+  <version>1.0</version>
+  <sdf version="1.6">model.sdf</sdf>
+
+  <author>
+  </author>
+
+  <description>
+  </description>
+</model>
