+++ buoy_description/models/mbari_wec/model.sdf
@@ -1,0 +1,211 @@
+<?xml version="1.0" ?>
+<sdf version="1.8">
+  <model name="MBARI_WEC">
+    <link name="Buoy">
+      <pose relative_to="__model__">0 0 0 0 0 0</pose>
+      <inertial>
+        <mass>1080</mass>
+        <inertia>
+          <ixx>10000</ixx>
+          <ixy>0</ixy>
+          <ixz>0</ixz>
+          <iyy>10000</iyy>
+          <iyz>0</iyz>
+          <izz>1000</izz>
+        </inertia>
+      </inertial>
+      <visual name="visual">
+        <geometry>
+          <mesh>
+            <uri>meshes/buoy_float.stl</uri>
+          </mesh>
+        </geometry>
+        <!--color-->
+        <material>
+          <ambient>1.0 1.0 0.0 1</ambient>
+          <diffuse>1.0 1.0 0.0 1</diffuse>
+          <specular>1.0 1.0 0.0 1</specular>
+        </material>
+      </visual>
+    </link>
+
+    <link name="PTO">
+      <pose relative_to="Buoy">0 0 0 0 0 0</pose>
+      <inertial>
+        <mass>605</mass>
+        <inertia>
+          <ixx>10000</ixx>
+          <ixy>0</ixy>
+          <ixz>0</ixz>
+          <iyy>10000</iyy>
+          <iyz>0</iyz>
+          <izz>1000</izz>
+        </inertia>
+      </inertial>
+      <visual name="visual">
+        <geometry>
+          <mesh>
+            <uri>meshes/pto.stl</uri>
+          </mesh>
+        </geometry>
+        <!--color-->
+        <material>
+          <ambient>.2 .2 0.2 0.5</ambient>
+          <diffuse>.2 .2 0.2 0.5</diffuse>
+          <specular>.2 .2 0.2 0.5</specular>
+        </material>
+      </visual>
+      <collision name="collision">
+        <geometry>
+          <mesh>
+            <uri>meshes/pto_collision.stl</uri>
+          </mesh>
+        </geometry>
+      </collision>
+    </link>
+
+    <link name="Piston">
+      <pose relative_to="PTO">0 0 -3.50066 0 0 0</pose>
+      <inertial>
+        <mass>48</mass>
+        <inertia>
+          <ixx>10000</ixx>
+          <ixy>0</ixy>
+          <ixz>0</ixz>
+          <iyy>10000</iyy>
+          <iyz>0</iyz>
+          <izz>1000</izz>
+        </inertia>
+      </inertial>
+      <visual name="visual">
+        <geometry>
+          <mesh>
+            <uri>meshes/rod_and_piston.stl</uri>
+          </mesh>
+        </geometry>
+        <!--color-->
+        <material>
+          <ambient>0.0 0.0 1.0 1</ambient>
+          <diffuse>0.0 0.0 1.0 1</diffuse>
+          <specular>0.0 0.0 1.0 1</specular>
+        </material>
+      </visual>
+    </link>
+
+    <link name="HeaveCone">
+      <pose relative_to="Piston">0 0 -10 0 0 0</pose>
+      <inertial>
+        <mass>1000</mass>
+        <inertia>
+          <ixx>10000</ixx>
+          <ixy>0</ixy>
+          <ixz>0</ixz>
+          <iyy>10000</iyy>
+          <iyz>0</iyz>
+          <izz>1000</izz>
+        </inertia>
+      </inertial>
+      <visual name="visual">
+        <geometry>
+          <mesh>
+            <uri>meshes/heave_cone.stl</uri>
+          </mesh>
+        </geometry>
+        <!--color-->
+        <material>
+          <ambient>0.1 0.1 .1 1</ambient>
+          <diffuse>0.1 0.1 .1 1</diffuse>
+          <specular>0.1 0.1 .1 1</specular>
+        </material>
+      </visual>
+    </link>
+      
+    <link name="Trefoil">
+      <pose relative_to="HeaveCone">0 0 0 0 0 0</pose>
+      <inertial>
+        <mass>10</mass>
+        <inertia>
+          <ixx>10000</ixx>
+          <ixy>0</ixy>
+          <ixz>0</ixz>
+          <iyy>10000</iyy>
+          <iyz>0</iyz>
+          <izz>1000</izz>
+        </inertia>
+      </inertial>
+      <visual name="visual">
+        <geometry>
+          <mesh>
+            <uri>meshes/trefoil.stl</uri>
+          </mesh>
+        </geometry>
+        <!--color-->
+        <material>
+          <ambient>0.1 0.1 .1 1</ambient>
+          <diffuse>0.1 0.1 .1 1</diffuse>
+          <specular>0.1 0.1 .1 1</specular>
+        </material>
+      </visual>
+    </link>
+
+    <joint name="WaterPlaneArea" type="prismatic">
+      <parent>world</parent>
+      <child>Buoy</child>
+      <pose>0 0 0 0 0 0</pose>
+      <axis>
+        <limit>
+          <lower>-2.0</lower>
+          <upper>2.0</upper>
+        </limit>
+        <xyz>0.0 0.0 1.0</xyz>
+        <dynamics>
+          <spring_stiffness>55000</spring_stiffness>
+          <spring_reference>0.0</spring_reference>
+          <damping>1000.0</damping>
+          <friction>0.0</friction>
+        </dynamics>
+      </axis>
+    </joint>
+
+    <joint name="Universal" type="ball">
+      <parent>Buoy</parent>
+      <child>PTO</child>
+      <provide_feedback>1</provide_feedback>
+      <pose>0.0 0.0 0.0 0 0 0</pose>
+    </joint>
+    
+    <joint name="TetherProxy" type="fixed">
+      <parent>Piston</parent>
+      <child>HeaveCone</child>
+      <provide_feedback>1</provide_feedback>
+      <pose>0.0 0.0 -10.0 0 0 0</pose>
+    </joint>
+
+    <joint name="HydraulicRam" type="prismatic">
+      <parent>PTO</parent>
+      <child>Piston</child>
+      <pose>0.0 0.0 0.0 0 0 0</pose>
+      <axis>
+        <limit>
+          <lower>0.0</lower>
+          <upper>2.03</upper>
+        </limit>
+        <xyz>0.0 0.0 1.0</xyz>
+      </axis>
+    </joint>
+
+    <joint name="TrefoilDoors" type="revolute">
+      <parent>HeaveCone</parent>
+      <child>Trefoil</child>
+      <provide_feedback>1</provide_feedback>
+      <pose>0.0 0.0 0.0 0 0 0</pose>
+      <axis>
+        <limit>
+          <lower>0.0</lower>
+          <upper>1.047</upper>
+        </limit>
+        <xyz>0.0 0.0 1.0</xyz>
+      </axis>
+    </joint>
+  </model>
+</sdf>
