--- buoy_gazebo/launch/buoy.launch.py
@@ -1,43 +1,0 @@
-# Copyright 2022 Open Source Robotics Foundation, Inc. and Monterey Bay Aquarium Research Institute
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""Launch Gazebo world with a buoy."""
-
-import os
-
-from ament_index_python.packages import get_package_share_directory
-from launch import LaunchDescription
-from launch.actions import DeclareLaunchArgument
-from launch.actions import IncludeLaunchDescription
-from launch.launch_description_sources import PythonLaunchDescriptionSource
-
-
-def generate_launch_description():
-
-    pkg_ros_ign_gazebo = get_package_share_directory('ros_ign_gazebo')
-    pkg_buoy_gazebo = get_package_share_directory('buoy_gazebo')
-
-    gazebo = IncludeLaunchDescription(
-        PythonLaunchDescriptionSource(
-            os.path.join(pkg_ros_ign_gazebo, 'launch', 'ign_gazebo.launch.py'),
-        ),
-    )
-
-    return LaunchDescription([
-        DeclareLaunchArgument(
-          'ign_args',
-          default_value=[os.path.join(pkg_buoy_gazebo, 'worlds', 'buoy_no_plugins.sdf'), ''],
-          description='Ignition Gazebo arguments'),
-        gazebo
-    ])
